requirements.txt
python-telegram-bot>=20.7
requests>=2.31.0
matplotlib>=3.7.0
gspread>=6.0.0
google-auth>=2.0.0
